[FEAT] #9 - 장바구니 추가 로직 구현 완료 및 테스트 데이터 확인

+++ app/src/main/java/com/seom/banchan/data/repository/CartRepositoryImpl.kt
@@ -10,7 +10,7 @@ class CartRepositoryImpl @Inject constructor(
 ) : CartRepository {
 
     // 장바구니에 메뉴 추가
-    override suspend fun addMenuToCart(menu: CartMenuModel): Result<Int> {
+    override suspend fun addMenuToCart(menu: CartMenuModel): Result<Long> {
         return cartDataSource.addCartItem(menu)
     }
 }
